requirements.txt
fastapi~=0.110.0
starlette~=0.36.3
uvicorn
httpx~=0.27.0
mysql-connector-python==8.3.0
pydantic==2.6.2
python-dotenv==1.0.1
openai==1.13.3
pytest~=8.0.2
requests